tests+coherence: reliable multinode runs — port isolation, orphan-proof teardown, and the two real bugs the noise was hiding
Multinode test runs failed intermittently from a tangle of environmental
interference and two genuine bugs the noise made impossible to attribute.

Environmental layer.  Tests interfered through two shared-state channels:
every local config bound the same port base, so a straggler from one test
(a rank still releasing its listen socket, or an orphan from a timed-out
run) poisoned every subsequent test; and a test killed by ctest's TIMEOUT
left its ranks alive — the launcher children's PDEATHSIG was SIGTERM,
which a hung rank's graceful shutdown path ignores.

 - Children now get PDEATHSIG(SIGKILL).  It only fires when the master
   died without running its cleanup (which already does waitpid → TERM →
   KILL), i.e. abnormally — there is nothing left to shut down gracefully.
 - Every multinode ctest gets its own default_ports base through the
   config env-override, allocated in tests/CMakeLists.txt; the harness
   allocates its own range per arts run.  The override must carry the
   same port COUNT as the cfg it replaces (the count doubles as the
   per-node parallel-connection count); 2n_io gets a two-port range.
 - All bases — including the stock configs, moved to 25000 — now sit
   BELOW the kernel ephemeral port range (32768-60999).  A base inside it
   collides at random with the source port of any outgoing connection on
   the host and the bind dies instantly; this was the true identity of a
   class of "random startup flakes".
 - The harness escalates timeouts with -k 1 and reaps leftovers by
   /proc/<pid>/exe match (comm truncates at 15 chars, pkill -x misses).

With the environment quiet, the two remaining intermittents proved real.

Coherence: signed writer_count ownership tests.  The commutative counter
is deliberately allowed to go transiently negative when an INVALIDATE
races ahead of its GRANT (two wires, multiple receiver threads).  Three
ownership tests still compared it as unsigned, treating negative as
"owner": the RW local fast path bumped it from -1 back to 0 — silently
cancelling the invalidation's decrement — and the acquire dispatch served
RO from a non-owned, stale buffer.  The cancelled decrement leaves a
surplus +1, the owner's last release reads 1 instead of 0, the transfer
never fires, and every queued RW acquirer on every rank is stranded —
reproduced at ~80% by the ownership-reorder stress under a two-receiver
config, the counter trace showing -1 → unlogged +1 → final rest=1.
Compare signed at all three sites (fast-path entry, RC and LRC is_owner):
zero or negative means not owner.  20/20 stress passes (1/6 before).

Test bug: route_table_remote_guid was written against a 2-node world —
every non-zero rank created the SAME pre-reserved DB guid and wired the
SAME slot of the rank-0 EDT, so at 3+ ranks the duplicate signals drove
the EDT's pending-dependency count from 0 to -1 and the EDT was lost (an
intermittent hang previously misread as a time-sync stall).  Reworked to
scale: each non-zero rank owns a distinct guid (homed on that rank),
creates exactly its own DB, signals exactly its own slot, and the reader
verifies every per-rank payload, aborting loudly on mismatch.  10/10 at
3n and 4n.

Result: RC multinode suite green three consecutive runs (79/79 ×3) plus
85/85 single-node, LRC 79/79 + 85/85, full correctness harness with zero
arts-side failures.  A test failure now means a bug, not weather.

+++ libs/src/core/coherence/lrc.c
@@ -48,7 +48,11 @@ void arts_handler_db_acquire(void *item, void *args) {
   struct arts_db_cache_s *cache = &db->cache;
   arts_edt_dep_t *dep = &((arts_edt_dep_t *)arts_get_depv(edt))[slot];
   arts_db_access_mode_t mode = dep->mode;
-  bool is_owner = (cache->writer_count > 0);
+  /* Signed: the commutative writer_count is transiently negative when an
+   * INVALIDATE races ahead of its GRANT (multi-receiver wire reorder) —
+   * negative means NOT owner.  An unsigned compare would treat it as owner
+   * and serve RO from a non-owned (stale) buffer. */
+  bool is_owner = ((int)cache->writer_count > 0);
 
   if (mode == DB_MODE_RO) {
     if (is_owner) { /* LRC RO predicate (only the owner holds the canonical
